refactored analysis service into new folder structure
