+++ backend/src/api/endpoints/search.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends
 
 import app.core.search.sdoc_search.sdoc_search as sdoc_search
 from api.dependencies import get_current_user
-from app.core.analysis.statistics import (
+from app.core.analysis.search_statistics.search_statistics import (
     compute_code_statistics,
     compute_keyword_statistics,
     compute_tag_statistics,
